requirements.txt
gspread >= 0.1.0, < 0.2.0
pillow
requests
PyYAML>=3.10
Jinja2>=2.7
